fix(rendering): use uint32 glyph indices in buildTextPageQuads so single-node text stops wrapping silently
buildTextPageQuads (TextLayout.ts) packed a single Text/BitmapText node's
own glyph placements into a Uint16Array index buffer, one layer above the
renderer-level index-cursor fix that already moved to Uint32 -- reachable
per node instead of per flush, as soon as one node alone exceeds 16384
visible glyphs on one atlas page. Same failure mode: silent index wrap,
no warning, wrong geometry.

Switch TextPageQuads.indices and buildTextPageQuads' index buffer to
Uint32Array, mirroring the prior renderer-level fix. No consumer needed
changes: both WebGl2TextRenderer and WebGpuTextRenderer already copy
batch.quads.indices element-by-element into their own (already Uint32)
staging buffers, dtype-agnostic.

Regenerated the TextPageQuads API doc page for the indices type change.

+++ src/rendering/text/TextLayout.ts
@@ -226,7 +226,7 @@ export function buildTextPageQuads(placements: readonly GlyphPlacement[]): TextP
     const n = pagePlacements.length;
     const vertices = new Float32Array(n * 8);
     const uvs = new Float32Array(n * 8);
-    const indices = new Uint16Array(n * 6);
+    const indices = new Uint32Array(n * 6);
 
     for (let i = 0; i < n; i++) {
       // In-bounds: i < n === pagePlacements.length.
